# docs/backend/openapi.v1.yaml
...
openapi: 3.1.0
info:
  title: LabelScan API
  version: v1
  description: Generated inventory of the currently implemented LabelScan API. Production disables the
    interactive OpenAPI and Swagger endpoints; this reviewed artifact defines pentest scope.
paths:
...
  /v1/ingestions/{ingestion_id}:
    get:
      summary: Get Ingestion
      operationId: get_ingestion_v1_ingestions__ingestion_id__get
      parameters:
      - name: ingestion_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          title: Ingestion Id
      - name: wait
        in: query
        required: false
        schema:
          type: number
          description: Long-poll hold in seconds (Tier 4), clamped to 25. Requires last_status; the response
            returns as soon as the status differs from last_status, or when the hold expires (with the
            current state).
          default: 0.0
          title: Wait
        description: Long-poll hold in seconds (Tier 4), clamped to 25. Requires last_status; the response
          returns as soon as the status differs from last_status, or when the hold expires (with the current
          state).
      - name: last_status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          description: The status the client last observed (long-poll baseline).
          title: Last Status
        description: The status the client last observed (long-poll baseline).
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionView'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
  /v1/extraction-runs/{run_id}:
    get:
      summary: Get Extraction Run
      operationId: get_extraction_run_v1_extraction_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionRunView'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    AuditEntry:
      properties:
        actor_id:
          type: string
          title: Actor Id
        action:
          type: string
          title: Action
        occurred_at:
          type: string
          title: Occurred At
      type: object
      required:
      - actor_id
      - action
      - occurred_at
      title: AuditEntry
...
    ExtractionRunView:
      properties:
        run_id:
          type: string
          title: Run Id
        ingestion_id:
          type: string
          title: Ingestion Id
        attempt_no:
          type: integer
          title: Attempt No
        outcome:
          type: string
          title: Outcome
        extractor_version:
          type: string
          title: Extractor Version
        prompt_version:
          type: string
          title: Prompt Version
        ocr_provider:
          type: string
          title: Ocr Provider
        llm_model:
          type: string
          title: Llm Model
        ocr_raw_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Ocr Raw Ref
        rule_set_version:
          type: string
          title: Rule Set Version
        created_at:
          type: string
          title: Created At
        fields:
          items:
            $ref: '#/components/schemas/FieldView'
          type: array
          title: Fields
        audit:
          items:
            $ref: '#/components/schemas/AuditEntry'
          type: array
          title: Audit
      type: object
      required:
      - run_id
      - ingestion_id
      - attempt_no
      - outcome
      - extractor_version
      - prompt_version
      - ocr_provider
      - llm_model
      - ocr_raw_ref
      - rule_set_version
      - created_at
      - fields
      - audit
      title: ExtractionRunView
    FieldView:
      properties:
        field_name:
          type: string
          title: Field Name
        value:
          anyOf:
          - {}
          - type: 'null'
          title: Value
        evidence:
          anyOf:
          - {}
          - type: 'null'
          title: Evidence
        provenance:
          anyOf:
          - {}
          - type: 'null'
          title: Provenance
        source_raw_artifact_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Raw Artifact Id
        validation_status:
          type: string
          title: Validation Status
        warnings:
          title: Warnings
        llm_confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Llm Confidence
        ocr_confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Ocr Confidence
        combined_confidence:
          type: number
          title: Combined Confidence
        confidence_band:
          type: string
          title: Confidence Band
        source:
          type: string
          title: Source
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - field_name
      - value
      - evidence
      - provenance
      - source_raw_artifact_id
      - validation_status
      - warnings
      - llm_confidence
      - ocr_confidence
      - combined_confidence
      - confidence_band
      - source
      - created_at
      title: FieldView
...
    IngestionView:
      properties:
        ingestion_id:
          type: string
          title: Ingestion Id
        status:
          type: string
          title: Status
        image_ref:
          type: string
          title: Image Ref
        checksum_sha256:
          type: string
          title: Checksum Sha256
        barcode_raw:
          anyOf:
          - type: string
          - type: 'null'
          title: Barcode Raw
        client_captured_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Captured At
        server_received_at:
          type: string
          title: Server Received At
        correlation_id:
          type: string
          title: Correlation Id
        trace_id:
          type: string
          title: Trace Id
        raw_artifacts:
          items:
            $ref: '#/components/schemas/RawArtifactView'
          type: array
          title: Raw Artifacts
        extraction_runs:
          items:
            $ref: '#/components/schemas/RunSummary'
          type: array
          title: Extraction Runs
        latest_fields:
          anyOf:
          - items:
              $ref: '#/components/schemas/FieldView'
            type: array
          - type: 'null'
          title: Latest Fields
        interim_fields:
          anyOf:
          - items:
              $ref: '#/components/schemas/InterimFieldView'
            type: array
          - type: 'null'
          title: Interim Fields
        recapture_required:
          type: boolean
          title: Recapture Required
          default: false
        recapture_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Recapture Reason
        audit:
          items:
            $ref: '#/components/schemas/AuditEntry'
          type: array
          title: Audit
      type: object
      required:
      - ingestion_id
      - status
      - image_ref
      - checksum_sha256
      - barcode_raw
      - client_captured_at
      - server_received_at
      - correlation_id
      - trace_id
      - raw_artifacts
      - extraction_runs
      - audit
      title: IngestionView
    InterimFieldView:
      properties:
        field_name:
          type: string
          title: Field Name
        value:
          title: Value
        source:
          type: string
          title: Source
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - field_name
      - value
      - source
      - created_at
      title: InterimFieldView
...
    RawArtifactView:
      properties:
        id:
          type: string
          title: Id
        artifact_kind:
          type: string
          title: Artifact Kind
        storage_ref:
          type: string
          title: Storage Ref
        checksum_sha256:
          type: string
          title: Checksum Sha256
        occurred_at:
          type: string
          title: Occurred At
      type: object
      required:
      - id
      - artifact_kind
      - storage_ref
      - checksum_sha256
      - occurred_at
      title: RawArtifactView
...
    RunSummary:
      properties:
        run_id:
          type: string
          title: Run Id
        attempt_no:
          type: integer
          title: Attempt No
        outcome:
          type: string
          title: Outcome
        extractor_version:
          type: string
          title: Extractor Version
        llm_model:
          type: string
          title: Llm Model
        ocr_provider:
          type: string
          title: Ocr Provider
        created_at:
          type: string
          title: Created At
        is_latest:
          type: boolean
          title: Is Latest
      type: object
      required:
      - run_id
      - attempt_no
      - outcome
      - extractor_version
      - llm_model
      - ocr_provider
      - created_at
      - is_latest
      title: RunSummary
...
    Problem:
      type: object
      description: Sanitized RFC 9457-style error envelope.
      required:
      - type
      - title
      - status
      - detail
      - error_code
      - timestamp
      - retriable
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        error_code:
          type: string
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
        timestamp:
          type: string
          format: date-time
        retriable:
          type: boolean
        errors:
          type: array
          items:
            type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute JWT with jti, sid, issuer, audience, tenant, role, and scopes
servers:
- url: /
  description: same-origin deployment